Refactored Performance into PerformanceHandler
